Rename 'mediaProperty' folder to 'properties'
