+++ app/avMeta/avMeta.cpp
@@ -1,5 +1,5 @@
 #include <AvTranscoder/file/InputFile.hpp>
-#include <AvTranscoder/mediaProperty/print.hpp>
+#include <AvTranscoder/properties/print.hpp>
 
 #include <AvTranscoder/progress/NoDisplayProgress.hpp>
