+++ src/AvTranscoder/file/InputFile.cpp
@@ -1,12 +1,12 @@
 #include "InputFile.hpp"
 
-#include <AvTranscoder/mediaProperty/util.hpp>
-#include <AvTranscoder/mediaProperty/VideoProperties.hpp>
-#include <AvTranscoder/mediaProperty/AudioProperties.hpp>
-#include <AvTranscoder/mediaProperty/DataProperties.hpp>
-#include <AvTranscoder/mediaProperty/SubtitleProperties.hpp>
-#include <AvTranscoder/mediaProperty/AttachementProperties.hpp>
-#include <AvTranscoder/mediaProperty/UnknownProperties.hpp>
+#include <AvTranscoder/properties/util.hpp>
+#include <AvTranscoder/properties/VideoProperties.hpp>
+#include <AvTranscoder/properties/AudioProperties.hpp>
+#include <AvTranscoder/properties/DataProperties.hpp>
+#include <AvTranscoder/properties/SubtitleProperties.hpp>
+#include <AvTranscoder/properties/AttachementProperties.hpp>
+#include <AvTranscoder/properties/UnknownProperties.hpp>
 
 extern "C" {
 #include <libavcodec/avcodec.h>
