+++ src/AvTranscoder/file/InputFile.hpp
@@ -5,7 +5,7 @@
 #include <AvTranscoder/file/util.hpp>
 #include <AvTranscoder/file/FormatContext.hpp>
 #include <AvTranscoder/stream/InputStream.hpp>
-#include <AvTranscoder/mediaProperty/FileProperties.hpp>
+#include <AvTranscoder/properties/FileProperties.hpp>
 #include <AvTranscoder/progress/IProgress.hpp>
 #include <AvTranscoder/profile/ProfileLoader.hpp>
