+++ src/AvTranscoder/properties/AudioProperties.cpp
