+++ src/AvTranscoder/properties/PixelProperties.cpp
