+++ src/AvTranscoder/properties/VideoProperties.cpp
