+++ src/AvTranscoder/reader/AudioReader.cpp
@@ -4,7 +4,7 @@
 #include <AvTranscoder/frame/AudioFrame.hpp>
 #include <AvTranscoder/transform/AudioTransform.hpp>
 #include <AvTranscoder/progress/NoDisplayProgress.hpp>
-#include <AvTranscoder/mediaProperty/print.hpp>
+#include <AvTranscoder/properties/print.hpp>
 
 namespace avtranscoder
 {
