+++ src/AvTranscoder/reader/VideoReader.cpp
@@ -4,7 +4,7 @@
 #include <AvTranscoder/frame/VideoFrame.hpp>
 #include <AvTranscoder/transform/VideoTransform.hpp>
 #include <AvTranscoder/progress/NoDisplayProgress.hpp>
-#include <AvTranscoder/mediaProperty/print.hpp>
+#include <AvTranscoder/properties/print.hpp>
 
 namespace avtranscoder
 {
